requirements.txt
sphinx_bootstrap_theme
numpydoc
sphinx_gallery
pywavelets
tqdm
sklearn
seaborn
datetime
scipy
mne
imbalanced-learn